# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for the Buyverse Platform",
    "version": "1.0.0",
    "title": "Buyverse API",
    "contact": {
      "email": "support@buyverse.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/verify-otp": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Verify OTP",
        "description": "Endpoint to verify OTP received by the user.",
        "operationId": "verifyOTP",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "OTP verification data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyOTP"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "VerifyOTP": {
      "type": "object",
      "required": [
        "email",
        "otp"
      ],
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "john.doe@example.com"
        },
        "otp": {
          "type": "string",
          "example": "123456"
        }
      }
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Operation completed successfully"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "An error occurred"
        },
        "error": {
          "type": "string",
          "example": "Detailed error message"
        }
      }
    },
...
  }
}